DEV_01 building up the matrix class

+++ src/test/java/com/craftsentient/craftmind/neuron/NeuronTest.java
@@ -13,25 +13,25 @@ public class NeuronTest {
     Double BIAS = 1.0;
 
     @BeforeEach
-    public void setUp(){
+    public void setUp() {
         inputs.add(1.0);
         weights.add(1.0);
     }
 
     @Test
-    public void generateOutputTest(){
+    public void generateOutputTest() {
         Neuron neuron = Neuron.builder()
                 .weights(weights)
                 .bias(BIAS)
                 .output(0.0)
                 .build();
-        Double output = neuron.generateOutput(inputs);
+        Double output = (Double)neuron.generateOutput(inputs);
 
-        Assertions.assertEquals(2, neuron.getOutput());
+        Assertions.assertEquals(2.0, neuron.getOutput());
     }
 
     @Test
-    public void createNeuronTest(){
+    public void createNeuronTest() {
         Neuron neuron = new Neuron(4, 2.0, 1, -1);
         Assertions.assertNotNull(neuron);
         Assertions.assertEquals(1, neuron.getMax());
@@ -40,7 +40,7 @@ public class NeuronTest {
     }
 
     @Test
-    public void generateNeuronTest(){
+    public void generateNeuronTest() {
         Neuron neuron = new Neuron(4, 1.0);
         System.out.println(neuron.getWeights());
         Assertions.assertNotNull(neuron);
